requirements.txt
asyncpg==0.21.0
discord.py==1.5.0
aiohttp==3.6.2

pre-commit==2.7.1
black==20.8b1